charge: fix pe2 charge stop and set ibus to 500mA
In factory image and PE2 charging up to 65,the charge will stop.
And the func mt6375_reset_pe_ta will set ibus to 500mA.
During the test,if the power consumption of phone is greater than
500mA,the battery will start to discharge.

So we need to restore ibus to the old value.

Change-Id: Ibc919d7318956f4a088e4e75a406a916981329dd
Signed-off-by: Guocheng Wang <wanggc6@motorola.com>
Reviewed-on: https://gerrit.mot.com/2686847
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Haijian Ma <mahj8@motorola.com>
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Reviewed-by: Lu Chai <chailu1@motorola.com>
Submit-Approved: Jira Key
Reviewed-on: https://gerrit.mot.com/2814695