gud: Workaround TCI inconsistency between REE and TEE.
It's found the Kernel TCI structure will not be processed with the desired
alignment(8 bytes) whereas in the secure world this will be handled. In
the event the TCI structure is not consistent between the two places. So
here in our own code we will workaround this for a little bit until we
find a good way to fix this.

Change-Id: Ife1646a192cbffc53b03d4f68ae6f574fd27cde3
Signed-off-by: quebs2 <quebs2@motorola.com>
Reviewed-on: https://gerrit.mot.com/2807504
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key