drivers/ccci: Support the modem CID secure authentication
Performing the secure authentication in Modem against the dangerous AT/Meta
commands are critical. Previously we used the approach to securely protect the
modem boot mode via the MTK special register(secure write only in ATF),
but here on mt6855, the register cannot be provided though.
Thus we would have to seek the alternative solution to maintain the signed
Motorola data for the auth. The idea is to use the shared memory carved out from
the LK and pass it over to modem for the ephemeral data management.

Change-Id: Ia0d957b62b17e004110a26638ba9c23debfab708
Reviewed-on: https://gerrit.mot.com/2832101
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Binsheng Que <quebs2@motorola.com>
Submit-Approved: Jira Key