mkp: Switch to use the kprobe to parse the kernel ro region
This is essentially needed because of the GKI restriction as well as the potential
compatible issues if we have dependencies on MTK mrdump driver. This is to decouple
the MKP from what Android debug info provided in Kernel6.1 and thus we can securely
lock the region without turning on the mrdump feature(which is dangerous on security).

Change-Id: I585542c669915e6283239eec2873f8833c317d55
Signed-off-by: quebs2 <quebs2@motorola.com>
Reviewed-on: https://gerrit.mot.com/2923003
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key