[ALPS09657841] iommu: modify the iova range reserve algorithm
[Description]
The previous iova range reserve algorithm would cause the range
that was expected to be reserved to not be reserved when the iova
bases were equal, so modify the iova range reserve algorithm to
avoid this problem.

[Test]
Check whether the iova range is correct.

MTK-Commit-Id: 5d62578e8dabd701eb18ed244e839359fc7a284f

Signed-off-by: ot908399 <ot_kelei.wang@mediatek.com>
CR-Id: ALPS09657841
Feature: iommu_kernel
Change-Id: Ibd8d688f8eb8a281a197bf297a8ecdb3fa5ba738
(cherry-pick from 0161ea50584d47eac912ebe3fcde009028109ce5)